+++ src/core/hist_grad_sum.cuh
@@ -13,21 +13,21 @@ __global__ void hist_sum_multi_node(
   SUM_T *dst_sum, unsigned *dst_count, const SUM_T *hist_sum_parent,
   const unsigned *hist_count_parent, const GRAD_T *__restrict__ values,
   const unsigned *__restrict__ parent_count_iter,
-  const unsigned *__restrict__ bin, const unsigned hist_size,
+  const unsigned short *__restrict__ bin, const unsigned hist_size,
   const unsigned end_bit, const int blocks_per_node, const bool use_trick);
 
 template <typename SUM_T, typename GRAD_T>
 __global__ void hist_sum(SUM_T *dst_sum, unsigned *dst_count,
                          const GRAD_T *__restrict__ values,
                          const unsigned *__restrict__ parent_count_iter,
-                         const unsigned *__restrict__ bin,
+                         const unsigned short *__restrict__ bin,
                          const unsigned end_bit, const size_t n);
 
 template <typename SUM_T, typename GRAD_T,
           int ITEMS_PER_THREAD = HIST_SUM_ITEMS_PER_THREAD>
 __global__ void hist_sum_node(SUM_T *dst_sum, unsigned *dst_count,
                               const GRAD_T *__restrict__ values,
-                              const unsigned *__restrict__ bin,
+                              const unsigned short *__restrict__ bin,
                               const unsigned end_bit, const unsigned segment,
                               const size_t n);
 
@@ -36,7 +36,7 @@ __global__ void hist_sum_dynamic(
   SUM_T *dst_sum, unsigned *dst_count, const SUM_T *hist_sum_parent,
   const unsigned *hist_count_parent, const GRAD_T *__restrict__ values,
   const unsigned *__restrict__ parent_count_iter,
-  const unsigned *__restrict__ bin, const unsigned hist_size,
+  const unsigned short *__restrict__ bin, const unsigned hist_size,
   const unsigned hist_size_bits, const bool use_trick, const size_t n);
 
 template <typename SUM_T>
