+++ src/core/hist_tree_grower.h
@@ -17,7 +17,7 @@ class HistTreeGrower : public BaseGrower<NODE_T, GRAD_T, SUM_T> {
   static void HistSum(SUM_T *sum, unsigned *bin_count,
                       const SUM_T *hist_sum_parent,
                       const unsigned *hist_count_parent, const GRAD_T *grad,
-                      const unsigned *node_size, const unsigned *fvalue,
+                      const unsigned *node_size, const unsigned short *fvalue,
                       const unsigned hist_size_bits, const unsigned hist_size,
                       const unsigned size, const bool use_trick,
                       cudaStream_t stream = 0);
@@ -26,19 +26,20 @@ class HistTreeGrower : public BaseGrower<NODE_T, GRAD_T, SUM_T> {
                             const SUM_T *hist_sum_parent,
                             const unsigned *hist_count_parent,
                             const GRAD_T *grad, const unsigned *node_size,
-                            const unsigned *fvalue,
+                            const unsigned short *fvalue,
                             const unsigned hist_size_bits,
                             const unsigned hist_size, const unsigned size,
                             const bool use_trick, cudaStream_t stream = 0);
 
   static void HistSumSingleNode(SUM_T *sum, unsigned *bin_count,
                                 const GRAD_T *grad, const unsigned *node_size,
-                                const unsigned *fvalue,
+                                const unsigned short *fvalue,
                                 const unsigned hist_size_bits,
                                 const unsigned size, cudaStream_t stream = 0);
 
   static void HistSumNaive(SUM_T *sum, unsigned *bin_count, const GRAD_T *grad,
-                           const unsigned *node_size, const unsigned *fvalue,
+                           const unsigned *node_size,
+                           const unsigned short *fvalue,
                            const unsigned hist_size, const unsigned size,
                            cudaStream_t stream = 0);
 
@@ -46,8 +47,8 @@ class HistTreeGrower : public BaseGrower<NODE_T, GRAD_T, SUM_T> {
   void ProcessDenseFeature(const device_vector<unsigned> &partitioning_index,
                            const device_vector<NODE_T> &row2Node,
                            const device_vector<GRAD_T> &grad_d,
-                           device_vector<unsigned> &fvalue_d,
-                           unsigned int *fvalue_h,
+                           device_vector<unsigned short> &fvalue_d,
+                           unsigned short *fvalue_h,
                            const device_vector<SUM_T> &parent_node_sum,
                            const device_vector<unsigned int> &parent_node_count,
                            const unsigned char fvalue_size,
@@ -58,8 +59,8 @@ class HistTreeGrower : public BaseGrower<NODE_T, GRAD_T, SUM_T> {
   template <typename NODE_VALUE_T>
   inline void ProcessCategoryFeature(
     const device_vector<NODE_T> &row2Node, const device_vector<GRAD_T> &grad_d,
-    const device_vector<unsigned int> &fvalue_d,
-    const host_vector<unsigned int> &fvalue_h,
+    const device_vector<unsigned short> &fvalue_d,
+    const host_vector<unsigned short> &fvalue_h,
     const device_vector<SUM_T> &parent_node_sum,
     const device_vector<unsigned int> &parent_node_count,
     const unsigned char category_size, const size_t level,
